requirements.txt
Flask
python-dotenv
requests
wave
sarvamai